requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch --index-url https://download.pytorch.org/whl/cpu
transformers
Pillow
timm
kornia
numpy
python-dotenv
